Remove LexAST and make Lexer a struct with an impl.
There's no use in exposing the LexAST to the outside world, when we can just
present a lexer instance, thus simplifying the API.
